Create a pytest_plugin.py file
Inspired by aiohttp, this allows other Python packages to
use the fixtures in their tests.

All that's required for pytest to detect the fixtures is
the following line in conftest.py:

pytest_plugins: list[str] = ["context_menu.pytest_plugin"]
